+++ src/main/java/com/dh/clinica/repository/DomicilioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DomicilioRepository extends JpaRepository<Domicilio, Integer> {
+public interface DomicilioRepository extends JpaRepository<Domicilio, Long> {
 
 }
